// IBT-Admin/src/types/settings.ts
export type SettingsKey = (typeof SETTINGS_KEYS)[keyof typeof SETTINGS_KEYS]

export type SettingEntity = {
  id: string
  key: SettingsKey
  value: unknown
  createdAt: string
  updatedAt: string
}